首页面写好了

+++ app/src/main/java/com/phc/neckrreferential/modle/domain/Categories.java
@@ -31,7 +31,7 @@ public class Categories {
         this.success = success;
     }
 
-    public int getCode() {
+    public long getCode() {
         return code;
     }
